Pregnancy_Chat_Bot/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
pypdf
langchain-text-splitters
python-dotenv
google-genai
chromadb